Moved _TimedInitMeta to utils and made it private
